# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Traveloop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ── Trips ──────────────────────────────────────────────────────────────────
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips for current user
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{tripId}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip by ID
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips/{tripId}/budget:
    get:
      operationId: getTripBudget
      tags: [trips]
      summary: Get budget breakdown for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripBudget"

  # ── Stops ──────────────────────────────────────────────────────────────────
  /trips/{tripId}/stops:
    get:
      operationId: listStops
      tags: [stops]
      summary: List stops for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of stops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripStop"
    post:
      operationId: createStop
      tags: [stops]
      summary: Add a stop to a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripStopInput"
      responses:
        "201":
          description: Created stop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripStop"
  /trips/{tripId}/stops/{stopId}:
    patch:
      operationId: updateStop
      tags: [stops]
      summary: Update a trip stop
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripStopUpdate"
      responses:
        "200":
          description: Updated stop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripStop"
    delete:
      operationId: deleteStop
      tags: [stops]
      summary: Delete a trip stop
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Stop Activities ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, name, email]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        language:
          type: string
        savedDestinations:
          type: array
          items:
            type: string
    UserProfileInput:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        language:
          type: string
        savedDestinations:
          type: array
          items:
            type: string

    # ── Trip ──────────────────────────────────────────────────────────────────
    Trip:
      type: object
      required: [id, userId, name, startDate, endDate, isPublic, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        coverPhotoUrl:
          type: ["string", "null"]
        isPublic:
          type: boolean
        totalBudget:
          type: ["number", "null"]
        createdAt:
          type: string
        stopCount:
          type: integer
    TripInput:
      type: object
      required: [name, startDate, endDate]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        coverPhotoUrl:
          type: string
        isPublic:
          type: boolean
        totalBudget:
          type: number
    TripUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        coverPhotoUrl:
          type: ["string", "null"]
        isPublic:
          type: boolean
        totalBudget:
          type: ["number", "null"]
    TripBudget:
      type: object
      required: [tripId, totalEstimated, totalBudget, breakdown, perDayCost, isOverBudget]
      properties:
        tripId:
          type: integer
        totalEstimated:
          type: number
        totalBudget:
          type: ["number", "null"]
        isOverBudget:
          type: boolean
        perDayCost:
          type: number
        breakdown:
          type: object
          required: [activities, accommodation, transport, meals]
          properties:
            activities:
              type: number
            accommodation:
              type: number
            transport:
              type: number
            meals:
              type: number

    # ── Stop ──────────────────────────────────────────────────────────────────
    TripStop:
      type: object
      required: [id, tripId, cityId, cityName, cityCountry, startDate, endDate, order]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        cityId:
          type: integer
        cityName:
          type: string
        cityCountry:
          type: string
        cityImageUrl:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        order:
          type: integer
        notes:
          type: ["string", "null"]
        activityCount:
          type: integer
    TripStopInput:
      type: object
      required: [cityId, startDate, endDate, order]
      properties:
        cityId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        order:
          type: integer
        notes:
          type: string
    TripStopUpdate:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        order:
          type: integer
        notes:
          type: ["string", "null"]

    # ── Activity ──────────────────────────────────────────────────────────────
...