# lib/api-zod/src/generated/types/tripUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TripUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  startDate?: string;
  endDate?: string;
  /** @nullable */
  coverPhotoUrl?: string | null;
  isPublic?: boolean;
  /** @nullable */
  totalBudget?: number | null;
}