# lib/api-zod/src/generated/types/userProfileInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileInput {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
  language?: string;
  savedDestinations?: string[];
}